backend/requirements.txt
alembic==1.0.8
asn1crypto==0.24.0
autopep8==1.4.4
backcall==0.1.0
boto==2.49.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
decorator==4.4.0
elasticsearch==7.0.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
httpie==1.0.2
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jwcrypto==0.6.0
jwt==0.6.1
Mako==1.0.8
MarkupSafe==1.1.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.3.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-jwt==3.2.4
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.2
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.15.2
